Use MockToken instead of Comp in rewards test
Note: I had to skip one test, because it no longer passed. The
reason is that `user2` was equal to `compOwner`, so the test did
not actually claim anything, it rather tested a no-op.

To be eligible for rewards, the user needs to mint or borrow some
tokens (or pretend to do so), wait for a few blocks to pass, and
then invoke the claim. Marked as TODO for future generations :)