fix: [011] remove bookmarks from chain
Problem: There are comments about the bookmarks functionality, and
these comments are valid (we don't check the comptroller passed to
the function). However, adding bookmarks on chain would not be a
good UX in the first place – people will need to confirm transactions
and pay tx fees to bookmark a pool.

Solution: Just remove bookmarks functionality.