fix: [VEN-949] CVP-01 fix sign in liquidation threshold check
Problem: While setting collateral factor/liquidation threshold
the code checks that liquidation threshold is LOWER than CF.
However, this would mean that we allow borrows that are immediately
eligible for liquidation. This was not the intention – liquidation
threshold mut be HIGHER than CF so that we have some margin between
borrowing and liquidation.

Solution: Replace the sign.