test: remove a failing test for irrelevant functionality
Problem: a test for getAccountLiquidity said the function should
revert if the account had not enabled any markets as collateral.
However, it should return 0, and there is a test that checks that
it returns 0. The test in question, in turn, checked that
liquidateAccount method failed if the market was not listed. We do
have a similar test in liquidateAccount test suite, and this test
has the corerct preconditions, there is no reason to have a duplicate.

Solution: Remove this test.