fix: prevent decimals underflow in exchange rate computation
Problem: The code fails for tokens with underlying decimals < 8.
Additionally, the code assumes that vTokens always have 8 decimals,
and although it's a correct assumption, we should have a protection
in case someone configures input.decimals to be a different number.

Solution: Change the order of operations and replace 8 with the
provided decimals value.