fix(lens): [VEN-1328][PLV-01] fix pending reward computation
Problem: In case rewardDistributor.initializeMarket(...) is called
after a user borrowed funds, the user's initial index would be zero,
while the global index would be nonzero. In this case, the lens
contract returned 0, while the correct value would account for the
reward accumulated from the initialization until the current block.

Solution: Use the logic similar to the one in supply reward
computation: if the user has an index of zero, and the global index
is nonzero, assume the user has the index equal to the one configured
at the beginning of the distribution. Assuming the user borrowed at
block A, later the market was initialized at block B, and the current
block is C, the user will receive the rewards from block B to block C.