refactor: [L-04] replace safeApprove with approve
Problem: safeApprove is deprecated and may revert a transaction in
case there's remaining allowance due to some reason, e.g. if the
allowance wasn't used in full in one of the previous transactions.
In some cases this can lead to DoS. It's not applicable to our
specific case since we always zero out the allowance, but it's
generally unexpected and confusing.

Solution: replace safeApprove with plain approve.