test: refactor token test helpers to make them more reusable
Problem: We have makeVToken utility function that can be used in
tests. However, it is not really reusable – it doesn't support
custom underlying assets, it doesn't use beacon-based deployment,
etc. As a result, we have to deploy VTokens either manually or
using addMarket in tests. Since addMarket will no longer deploy
VToken contracts, the need for a reusable utility function is even
higher.

Solution: Change makeVToken to accept all the parameters the
VToken initializer receives, change deployment to beacon-based,
and fake all the contracts that are not provided to simplify the
tests.