chore: decrease maxLoopsLimit to 100
Decreasing the limit from 150 to 100 would allow us to be more
flexible in the loop contents. We can safely _increase_ the loops
limit later, but we can't decrease it, so starting with a more
conservative value makes sense. On the other hand, setting 50 as
the limit would potentially block innocent initiatives like adding
more markets to the pool.