feat: use core pool behavior for jump rate multiplier
Problem: The new IL JumpRateModel is based on Compound's updated
version. The problem with that is that the multiplier is divided
by kink in this updated version. This is counterintuitive and
inconsistend with the core pool behavior.

Solution: Use the classic computation, i.e. not divide the slope
by kink.