feat: avoid locking on failed transfers to previous bidders
Problem: Shortfall contract transfers tokens to previous bidders
on placeBid. This might lock the auction if the receiver can not
receive a transfer due to some reason, e.g. if the receiver is
blacklisted in the token contract.

Solution: Track the tokens owed to users due to failed transfer
attempts.