fix: [RFR-01][RFR-03] unify the meaning of minAmountToConvert
Problem: With the evolution of the code, the meaning of the
conversion threshold started to change. As a result, we ended up
comparing the amount out (in tokens), balance of underlying *
oracle price (in dollars), with minAmountToConvert. While in most
cases these values would be close to each other as long as some
stablecoin is used for the threshold, it might be problematic
to set this threshold correctly, considering the potential slippage.

Solution: Expect amountOutMin * base price to be greater than
minAmountToConvert. Adjust the failing test accordingly.