chore: use incremental indexes for rewards distributors
Problem: We're going to have several rewards distributors with the
same reward token in the same pool. This would lead to collisions
in deployment names.

Solution: Change deployment postfix to "_{poolId}_{index}", so that
when a new rewards distributor is appended to the rewards array,
it gets deployed regardless of whether reward token collides.