fix: revert on approval failures
Problem: OpenZeppelin safeApprove implementation is deprecated, so
its usage was removed from the RiskFund contract in favor of plain
approve call. However, this is not the correct way to fix the issue
since the updated version of the code doesn't handle the case when
tokens return false to indicate an error.

Solution: Use a custom library similar to Uniswap's TransferHelpers
to make an approve call supporting:
  * false to indicate failure
  * revert to indicate failure
  * non-compliant tokens that don't return any value