refactor: remove Compound's scenario tests
Problem: We've inherited a scenario test suite from Compound that
was useful in initial phases of development. However, these tests
are hard to maintain due to outdated toolset and a custom DSL used
in these tests. Scenario tests were not maintained, a separate set
of integration tests were implemented instead. There is no reason
to keep this folder in the repo.

Solution: Remove scenario tests.