fix: use Venus-style interest rate multiplier for bsctestnet models
Problem: Some of the testnet interest rate models use an old
Compound-style definition for interest rate multiplier, which
is multiplierPerBlock = multiplierPerYear / (blocksPerYear * kink).
The current Venus rate models do not divide the multiplier by
kink.

Solution: To avoid redeploying the interest rate models, use the
Venus definition of "multiplierPerYear" while keeping the current
contracts. The old contracts will be phased out and it is unlikely
we try to update the parameters since the common practice is to
redeploy the rate model contracts every time we change them.