refactor: make "time-based" a property of the network
Problem: Currently whether the deployment is time-based is defined
by the environment variable. This is error-prone and requires us to
support corner cases like having both blocksPerYear configured and
timeBased flag set to true.

Solution: Make "time-based" a property of a particular network.
For hardhat, support both modes based on the environment variable.