refactor: stop using BSC_BLOCKS_PER_YEAR in tests
Problem: Some of the tests rely on BSC_BLOCKS_PER_YEAR to be a
constant value, e.g. rewards and interests in tests are computed
under this assumption. Adjusting test cases for a different number
of blocks per year every time this value is changed is an unnecessary
overkill, since the logic itself doesn't depend on a particular
value assumed by the test suite.

Solution: Introduce a separate constant, DEFAULT_BLOCKS_PER_YEAR,
that should be used in tests going forward.